On error connecting, don't send QUIT command.
If createClient() can't connect to the redis server, it retries
on a loop.  It also (by default) queues up commands in an offline queue,
to be sent when connected.

On error connecting, this module was sending a QUIT command by calling .quit()
on the node-redis client.  All this did was queue up the command in the offline
queue - which evidently makes no sense at all.  This also resulted in an uncaught
exception, which crashed our app, because there was no longer a listener on the
client's 'error' event.

Instead, if we can't connect initially, we actually, genuinely kill the connection
by calling .end() on the node-redis client.

Subsequent errors (after calling the pool-redis getClient callback) will be emitted
and should be handled by apps.